requirements.txt
django
djangorestframework
gunicorn
requests
pytest
